Refactor regions_combine benchmarks into ugrid subdir.
