+++ benchmarks/benchmarks/experimental/ugrid/__init__.py
@@ -14,8 +14,8 @@ import numpy as np
 
 from iris.experimental import ugrid
 
-from .. import ARTIFICIAL_DIM_SIZE, disable_repeat_between_setup
-from ..generate_data.stock import sample_mesh
+from ... import ARTIFICIAL_DIM_SIZE, disable_repeat_between_setup
+from ...generate_data.stock import sample_mesh
 
 
 class UGridCommon:
